refactoring to avoid code duplication. Renamed the sampler
